refactor: split the CLI entry point from the server module
src/index.ts becomes a bootstrap that constructs and runs ApexLogServer.
The server moves to src/server.ts, which no longer has import side
effects, so tests can import it without starting a server. The bin path
(dist/index.js) is unchanged.
